Fix minimum lines algorithm
This uses a maximal match bipartite graph algorithm to get the minimum lines to cover the zeros (by converting the maximal match to a minimum vertex cover). TODO is to fix how zeros are selected. Sometimes it will get stuck selecting the wrong one which will lead to some columns not having a match. The current version sorts by number of zeros available, choosing the zeros with the least options first, for example:

```
   ·   ·   ·   ·   ·   0   0   0
   ·   ·   ·   ·   0   0   0   0
   ·   ·   ·   ·   ·   0   0   0
   0   ·   ·   ·   0   0   0   0
   ·   ·   ·   ·   0   0   0   0
   ·   0   0   0   ·   ·   ·   ·
   0   ·   0   ·   ·   ·   ·   ·
   ·   0   ·   0   ·   0   0   0
```

```
   ·   ·   ·   ·   ·   0   0   0
   ·   ·   ·   ·   0   0   0   0
   ·   ·   ·   ·   ·   0   0   0
   ⨯   ·   ·   ·   0   0   0   0
   ·   ·   ·   ·   0   0   0   0
   ·   0   0   0   ·   ·   ·   ·
   ✓   ·   ⨯   ·   ·   ·   ·   ·
   ·   0   ·   0   ·   0   0   0
```

```
   ·   ·   ·   ·   ·   0   0   0
   ·   ·   ·   ·   0   0   0   0
   ·   ·   ·   ·   ·   0   0   0
   ⨯   ·   ·   ·   0   0   0   0
   ·   ·   ·   ·   0   0   0   0
   ·   ⨯   ✓   ⨯   ·   ·   ·   ·
   ✓   ·   ⨯   ·   ·   ·   ·   ·
   ·   0   ·   0   ·   0   0   0
```

```
   ·   ·   ·   ·   ·   0   0   0
   ·   ·   ·   ·   0   0   0   0
   ·   ·   ·   ·   ·   0   0   0
   ⨯   ·   ·   ·   0   0   0   0
   ·   ·   ·   ·   0   0   0   0
   ·   ⨯   ✓   ⨯   ·   ·   ·   ·
   ✓   ·   ⨯   ·   ·   ·   ·   ·
   ·   ✓   ·   ⨯   ·   ⨯   ⨯   ⨯
```

Now column 3 has no options. Oops.