Fix zero selection, make transpose be in place
This fixes the zero selection by thinking of it as another bipartite graph – the maximum matching of the bipartite graph is the correct selection. This nextly also makes the transpose function change the matrix in place to make it a bit faster (no re-allocating). Next is to work on the battery aware version to make it actually better in some sense.